Hi joel, here are the many changes I'd made to the logic, structure, and setup of the ultrasonic range sensor class. I left it intentionally messy and in that sense, half broken so you could see where the code you left fits in, but there is a much more streamlined and simplified way to do what I did and I hope you can see and understand where I'm coming from.
Logic things range from checking if a user set value was true or false to determine if sensor was valid or not to not actually setting, or making an easy to have the sensor enabled when the idea was to have it running all the time.

Design/structure things continue off that last point and include the naming of methods like enable(state) instead of enable() and disable() and the fact that if the sensor was called for a single measurement without doing all of the enabling setup or not, the class failed to simplify the usage of the sensor, requiring that enable still be called and not reducing the code on the end programmer's end.

Functionality things go from leaving an unnecessary thread in the background running (in the original revision, i'm assuming. the thread was completely disabled in this revision) to where I implemented a basic average across a period of time mechanism so you can poll the sensor and retrieve the average distance. <- that's actually not going to be as useful as it may sound because I didn't implement a way to remove outliers in a dataset, and it's a waste of processing power. It could be more useful if it removed outlier data, but even then calling an instantaneous point is likely going to result in better data.

-adam li